Reorganize agent_r1 trainer modules
