+++ agent_r1/reward_loop/__init__.py
@@ -1,0 +1,1 @@
+"""Reward loop package."""
