+++ agent_r1/trainer/__init__.py
@@ -1,0 +1,1 @@
+"""Trainer package."""
