+++ agent_r1/trainer/ppo/__init__.py
@@ -1,0 +1,1 @@
+"""PPO trainer package."""
